+++ R/lenna.R
@@ -1,7 +1,0 @@
-#' Lenna Image
-#'
-#' A dataset containing Lena image.
-#'
-#' @format A numeric 512 x 512 matrix.
-#' @usage data(lenna)
-"lenna"